requirements.txt
langgraph>=0.0.30
langchain>=0.1.10
langchain-openai>=0.1.0
langchain-google-genai>=0.1.0
google-cloud-bigquery>=3.14.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-generativeai>=0.5.4
db-dtypes>=1.1.0
pandas>=2.1.0
plotly>=5.17.0
graphviz>=0.20.3
kaleido>=0.2.1
typer>=0.9.0
rich>=13.6.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.11.0
pytest>=7.4.0
ruff>=0.5.0

